Added JWT usage
All endpoints now require a JWT to be used. A missing token results in a
401 response
The classify and regex endpoints now have further validation using JWT. This
consists of the following:
 - For the classify endpoint:
    - Checks the body of the request and returns 403 if the JWT does not
    contain access to the domain specified.
    - Checks the amount of contents sent and returns 413 if the number
    exceeds the limit specified in the JWT.
 - For the regex endpoint:
    - Checks the domain of the request and returns 403 if the JWT does
    not contain access to the domain specified.
    - Checks the request method and returns 403 if the JWT does not
    contain permissions to use that request method.

Added Permissions class to take looks at and generate permission values given in
the JWT.

Commented bean creation in WebSecurityConfig. As this is now realized
through a OncePerRequestFilter this bean is not used anymore it seems.